Updated sample and readme file
