+++ SQLWithChart/MainWindow.xaml
@@ -2,11 +2,12 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:chart="clr-namespace:Syncfusion.UI.Xaml.Charts;assembly=Syncfusion.SfChart.WPF"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:SQLWithChart"
-        xmlns:chart="http://schemas.syncfusion.com/wpf"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
+
     <Grid>
 
         <Grid.DataContext>
@@ -24,9 +25,10 @@
             </chart:SfChart.SecondaryAxis>
 
             <chart:ScatterSeries ItemsSource="{Binding DataTable}"
-                                 XBindingPath="xval"
-                                YBindingPath="yval"/>
+                              XBindingPath="xval"
+                             YBindingPath="yval"/>
         </chart:SfChart>
 
     </Grid>
+
 </Window>
