+++ SQLWithChart/MainWindow.xaml.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Text;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
@@ -25,42 +19,6 @@ namespace SQLWithChart
         public MainWindow()
         {
             InitializeComponent();
-        }
-    }
-
-    public class ViewModel
-    {
-        public ViewModel()
-        {
-            try
-            {
-                SqlConnection thisConnection = new SqlConnection(ConnectionString);
-                thisConnection.Open();
-                string Get_Data = "SELECT * FROM ChartData";
-                SqlCommand cmd = thisConnection.CreateCommand();
-                cmd.CommandText = Get_Data;
-                SqlDataAdapter sda = new SqlDataAdapter(cmd);
-                DataSet ds = new DataSet();
-                sda.Fill(ds);
-                var table = ds.Tables[0];
-                this.DataTable = table;
-            }
-            catch
-            {
-                MessageBox.Show("DataBase Error");
-            }
-        }
-
-        public object DataTable { get; set; }
-
-        public static string ConnectionString
-        {
-            get
-            {
-                string currentDir = System.Environment.CurrentDirectory;
-                currentDir = currentDir.Substring(0, currentDir.Length - 10) + "\\LocalDataBase";
-                return @"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename=" + currentDir + @"\SeriesItemsSource.mdf;Integrated Security=True";
-            }
         }
     }
 }
